remove Windows support
A few attempts I tried did not work. Windows is always special and
there's something special with adding `lcov` to `PATH` or
whatever they have on Windows.

There should be Windows support but I need more time for that. So
for now only Linux and macOS runners should be supported.